actually split my classes across folders
