+++ src/main/java/doot/Doot.java
@@ -1,15 +1,6 @@
 package doot;
 
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.File;
-import java.time.format.DateTimeParseException;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.util.Scanner;
-import java.util.ArrayList;
-import java.util.List;
 
 import static doot.Logo.LOGO;
 
@@ -26,6 +17,7 @@ public class Doot {
             super(message);
         }
     }
+
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("________________________________________________________________________________________________________________________\n" +
@@ -48,544 +40,5 @@ public class Doot {
                 ________________________________________________________________________________________________________________________
                 """);
 
-    }
-
-    //class for wrapping the nice lines around text to be printed
-    public static class  UI {
-
-        //wraps text with underscores then print
-        //@param message the message to be printed
-        public static void showMessage(String message) {
-            System.out.println("________________________________________________________________________________________________________________________\n" +
-                    message +
-                    "\n________________________________________________________________________________________________________________________\n");        }
-    }
-
-
-    //This class is for interpreting what the user means. It contains a Tasklist so it can edit it
-    public static class Parser {
-        private TaskList list;
-
-        public Parser(TaskList list) {
-            this.list = list;
-        }
-
-        //the main method, takes in the user input and converts it to commands
-        //@userInput the input by the user
-        public void handleCommand(String userInput) {
-            if (userInput.isEmpty()) {
-                UI.showMessage("say something I'm giving up on you");
-            } else if (userInput.equals("thank mr skeltal")) {
-                UI.showMessage("good bones and calcium will come to you");
-            } else if (userInput.equals("doot doot")) {
-                UI.showMessage("doot doot");
-            } else if (userInput.equals("list")) {
-                UI.showMessage(list.returnList());
-            } else if (Parser.isMark(userInput)) {
-                handleMark(userInput);
-            } else if (Parser.isUnMark(userInput)) {
-                handleUnMark(userInput);
-            } else if (userInput.startsWith("delete ")) {
-                handleDelete(userInput);
-            } else if (userInput.equals("listData")) {
-                UI.showMessage(list.listData());
-            } else if (userInput.equals("savedata")) {
-                saveList();
-            } else if (userInput.startsWith("find ")) {
-                handleFind(userInput);
-            } else {
-                addTask(userInput);
-            }
-        }
-
-        //called by the parser when the user enters something starting with mark. It sets the task in TaskList to be marked
-        //@param userInput the entire userinput is fed back in
-        private void handleMark(String userInput) {
-            int index = Integer.parseInt(userInput.split(" ")[1]) - 1;
-            if (isValidIndex(index)) {
-                list.mark(index);
-                UI.showMessage("doot doot\n\n" + list.returnList());
-                saveList();
-            } else {
-                UI.showMessage("too big/too small number");
-            }
-        }
-
-        //sets the corresponding task as not completed
-        //@param userInput the user Input
-        private void handleUnMark(String userInput) {
-            int index = Integer.parseInt(userInput.split(" ")[1]) - 1;
-            if (isValidIndex(index)) {
-                list.unMark(index);
-                UI.showMessage("noot noot\n\n" + list.returnList());
-                saveList();
-            } else {
-                UI.showMessage("too big/too small number");
-            }
-        }
-
-        //It handles deletes
-        private void handleDelete(String userInput) {
-            int index = Integer.parseInt(userInput.split(" ")[1]) - 1;
-            if (isValidIndex(index)) {
-                Task removed = list.getTask(index);
-                list.removeTask(index);
-                UI.showMessage("calcium for you\n   removed " + removed.getDetails() + "\n" + list.size() + " more to do\n");
-                saveList();
-            } else {
-                UI.showMessage("too big/too small number");
-            }
-        }
-
-        //it adds tasks to the tasklist as according to the userinput
-        private void addTask(String userInput) {
-            try {
-                list.addTask(userInput);
-                saveList();
-            } catch (InvalidFormatException e) {
-                UI.showMessage(e.toString());
-            }
-        }
-
-        //called after every command that edits the list, it saves the list using the Storage class
-        private void saveList() {
-            try {
-                Storage.saveList(list);
-            } catch (IOException e) {
-                UI.showMessage(e.toString());
-            }
-        }
-
-        //determines if the index is valid for commands like delete or mark, like if there are only 3 tasks the command
-        //unmark 4 will return a false
-        private boolean isValidIndex(int index) {
-            return index >= 0 && index < list.size();
-        }
-
-        //determines if the command is a mark command, and if it is valid
-        public static boolean isMark(String str) {
-            if (str.startsWith("mark ") || str.startsWith("Mark ")) {
-                String[] arr = str.split(" ");
-                if (arr.length == 2) {
-                    try {
-                        Integer.parseInt(arr[1]);
-                        return true;
-                    } catch (NumberFormatException e) {
-                        return false;
-                    }
-                }
-            }
-            return false;
-        }
-
-        //determines if this is an unmark command, and checks if it is valid
-        public static boolean isUnMark(String str) {
-            if (str.startsWith("unmark ") || str.startsWith("Unmark ")) {
-                String[] arr = str.split(" ");
-                if (arr.length == 2) {
-                    try {
-                        Integer.parseInt(arr[1]);
-                        return true;
-                    } catch (NumberFormatException e) {
-                        return false;
-                    }
-                }
-            }
-            return false;
-        }
-
-        //just runs the searchWord() method for TaskList
-        public void handleFind(String str) {
-            String text = str.substring(5).strip();
-            if (text.equals("")) {
-                UI.showMessage("put something after the find");
-            } else {
-                String found = list.searchWord(text);
-                if (found.equals("")) {
-                    UI.showMessage("nothing found");
-                } else {
-                    UI.showMessage(found);
-                }
-            }
-        }
-    }
-
-    //the main structure that contains the tasks. Contains methods to modify the tasks' statuses
-    public static class TaskList {
-        private List<Task> arr;
-        public TaskList() {
-            arr = new ArrayList<>();
-        }
-
-        //creates a new task from the userinput str, adds it to the list, and creates a message
-        public void addTask(String str) throws InvalidFormatException {
-            Task task = Task.makeTask(str);
-            arr.add(task);
-            UI.showMessage("task added\n   " + task.getDetails() + "\nyou now have " + arr.size() + " tasks in the list");
-        }
-
-        //loads tasks by running the strings needed from the file.
-        //@param File f is the file which contains the past commands used to make the list
-        public void loadTask(File f) throws FileNotFoundException, InvalidFormatException {
-            Scanner s = new Scanner(f);
-            int count = 0;
-            while (s.hasNext()) {
-                String line = s.nextLine();
-                if (line.startsWith("d ")) {
-                    line = line.substring(2);
-                    Task task = Task.makeTask(line);
-                    arr.add(task);
-                    this.mark(count);
-                } else {
-                    Task task = Task.makeTask(line);
-                    arr.add(task);
-                }
-                count += 1;
-            }
-        }
-
-        //returns the entire list, in the format used for printing as a list when called by the user
-        public String returnList() {
-            int count = 1;
-            StringBuilder list = new StringBuilder();
-            for (Task task : arr) {
-                list.append(count).append(". ").append(task.getDetails()).append("\n");
-                count ++;
-            }
-            return list.toString().trim();
-        }
-
-        //used for saving. This assembles the creationString() of all tasks and seperates them in their own lines
-        public String listData() {
-            StringBuilder list = new StringBuilder();
-            for (Task task : arr) {
-                list.append(task.creationString()).append("\n");
-            }
-            return list.toString().trim();
-        }
-
-        //gets the size of the list
-        public int size() {
-            return arr.size();
-        }
-
-        //marks the task in the list
-        public void mark(int num) {
-            arr.get(num).setDone();
-        }
-
-        //unmarks the task in the list
-        public void unMark(int num) {
-            arr.get(num).setUndone();
-        }
-
-        //returns the corresponding task
-        public Task getTask(int num) {
-            return arr.get(num);
-        }
-
-        //deletes the task
-        public void removeTask(int num) {
-            arr.remove(num);
-        }
-
-        //returns the details of all tasks that has the str within their description
-        public String searchWord(String str) {
-            StringBuilder list = new StringBuilder();
-            for (int i = 0; i < this.size(); i++) {
-                if (this.getTask(i).isSubstring(str)) {
-                    list.append(this.getTask(i).getDetails()).append("\n");
-                }
-            }
-            return list.toString().strip();
-        }
-    }
-
-    public enum Type {T, D, E}
-
-    //the class for storing information on individual tasks.
-    //contains state, description, type. This is an abstract class so that the more specific task classes can inherit
-    public abstract static class Task {
-        protected String description;
-        protected boolean isDone;
-
-        protected Type type;
-
-        public Task(String description) {
-            this.description = description;
-            this.isDone = false;
-        }
-
-        //static class used to make a new class, then returns said class
-        public static Task makeTask(String str) throws InvalidFormatException{
-            String taskDescription = str.substring(str.indexOf(" ") + 1).trim();
-            if (taskDescription.isEmpty()) {
-                throw new InvalidFormatException("tasj must have description at back nothing added doot doot");
-            }
-
-            if (str.startsWith("todo ")) {
-                return new todoTask(str.substring(str.indexOf(" ") + 1));
-            } else if (str.startsWith("deadline ")) {
-                if (!str.contains("/by")) {
-                    throw new InvalidFormatException("deadline needs /by weak bone man");
-                }
-                String deadline = str.substring(str.indexOf("/by") + 4);
-                if (deadline.equals("-1") || !str.contains("/by")) {
-                    throw new InvalidFormatException("pls /by correctly");
-                }
-                if (str.indexOf("/by") - 2 - str.indexOf(" ") < 1) {
-                    throw new InvalidFormatException("u need sumting between deadline and /by");
-                }
-                return new DeadlineTask(str.substring(str.indexOf(" ") + 1, str.indexOf("/by") - 1), deadline);
-            } else if (str.startsWith("event ")) {
-                if (!str.contains("/from") || !str.contains("/to")) {
-                    throw new InvalidFormatException("need /from and /to keyword for event doot doot");
-                }
-
-                if (str.indexOf("/from") - 2 - str.indexOf(" ") < 1) {
-                    throw new InvalidFormatException("u need sumting between event and /from");
-                }
-
-                if (str.indexOf("/from") > str.indexOf("/to")) {
-                    throw new InvalidFormatException("/from before /to u dum");
-                }
-
-                if (str.indexOf("/from") + 6 > str.indexOf("/to") - 1) {
-                    throw new InvalidFormatException("oop you fuked up go put something between /from and /to");
-                }
-
-                if (str.length() <= str.indexOf("/to") + 3) {
-                    throw new InvalidFormatException("doot doot go put something behind /to space works also ");
-                }
-
-                String start = str.substring(str.indexOf("/from") + 6, str.indexOf("/to") - 1);
-                String end = str.substring(str.indexOf("/to") + 4);
-                if (start.equals("-1") || end.equals("-1")) {
-                    throw new InvalidFormatException("/from /to formatting wrong bet you have osteoporosis");
-                }
-                return new EventTask(str.substring(str.indexOf(" ") + 1, str.indexOf("/from") - 1),
-                        start,
-                        end);
-            } else {
-                throw new InvalidFormatException("dumbass");
-            }
-        }
-
-        //checks if the str is a substring of the description
-        public boolean isSubstring(String str) {
-            return this.description.contains(str);
-        }
-
-        public abstract String getType();
-
-        //contains the line needed to create a copy of this task
-        public abstract String creationString();
-
-        //returns a string in the format to be passed to the UI to be printed for the user to see
-        public String getDetails() {
-            return this.getStatusIcon() + " " + this.getDescription();
-        }
-
-        //returns the status of the task in a visually nicer manner
-        public String getStatusIcon() {
-            return "[" + (isDone ? "X" : " ") + "]"; // mark done task with X
-        }
-
-        //gets description
-        public String getDescription() {
-            return description;
-        }
-
-        //set task as complete
-        public void setDone() {
-            this.isDone = true;
-        }
-
-        //set task as no complete
-        public void setUndone() {
-            this.isDone = false;
-        }
-
-        //returns if the task is complete
-        public boolean isDone() {
-            return isDone;
-        }
-    }
-
-    //the subclass of task, specifically for the todo tasks
-    public static class todoTask extends Task {
-        public todoTask(String description) {
-            super(description);
-            type = Type.T;
-        }
-
-        //determines type of task, like T for todo
-        @Override
-        public String getType() {
-            return "[" +type.name() +"]";
-        }
-
-        //returns details for todo objects
-        @Override
-        public String getDetails() {
-            return this.getType() + this.getStatusIcon() + " " + this.getDescription();
-        }
-
-        //the string necessary to make this object again
-        @Override
-        public String creationString() {
-            StringBuilder list = new StringBuilder();
-            if (this.isDone) {
-                list.append("d ");
-            }
-            list.append("todo ").append(this.getDescription());
-
-            return list.toString();
-        }
-    }
-
-    //subclass of task, deadlinetask is the class for deadline tasks(wow)
-    public static class DeadlineTask extends Task {
-        protected String deadline;
-        protected LocalDate datedeadline;
-        public DeadlineTask(String description, String deadline) {
-            super(description);
-            this.type = Type.D;
-            this.deadline = deadline;
-            for (String format : formats) {
-                try {
-                    this.datedeadline = LocalDate.from(LocalDate.parse(deadline, DateTimeFormatter.ofPattern(format)).atStartOfDay());
-                } catch (DateTimeParseException ignored) {
-                }
-            }
-        }
-
-        @Override
-        public String getType() {
-            return "[" +type.name() +"]";
-        }
-
-        //deadline tasks include information on the actual deadline. This method converts the localtime to a string format
-        //if there isnt a localtime, it just returns the string instead
-        public String getDeadline() {
-            return !(datedeadline == null) ? datedeadline.format(DateTimeFormatter.ofPattern("dd MMM, yyyy")) : deadline;
-        }
-
-        @Override
-        public String getDetails() {
-            return this.getType() + this.getStatusIcon() + " " + this.getDescription() + " (by: " + getDeadline() + ")";
-        }
-
-        @Override
-        public String creationString() {
-            StringBuilder list = new StringBuilder();
-            if (this.isDone) {
-                list.append("d ");
-            }
-            list.append("deadline ").append(this.getDescription()).append(" /by ").append(deadline);
-
-            return list.toString();
-        }
-    }
-    public static class EventTask extends Task {
-        protected String start;
-        protected String end;
-
-        protected LocalDate datestart;
-
-        protected LocalDate dateend;
-
-        public EventTask(String description, String start, String end) {
-            super(description);
-            this.type = Type.E;
-            this.start = start;
-            this.end = end;
-
-
-
-            for (String format : formats) {
-                try {
-                    this.datestart = LocalDate.from(LocalDate.parse(start, DateTimeFormatter.ofPattern(format)).atStartOfDay());
-                } catch (DateTimeParseException ignored) {
-                }
-            }
-            for (String format : formats) {
-                try {
-                    this.dateend = LocalDate.from(LocalDate.parse(end, DateTimeFormatter.ofPattern(format)).atStartOfDay());
-                } catch (DateTimeParseException ignored) {
-                }
-            }
-        }
-
-        @Override
-        public String getType() {
-            return "[" +type.name() +"]";
-        }
-
-        @Override
-        public String getDetails() {
-            return this.getType() + this.getStatusIcon() + " " + this.getDescription() + " (from: " + getStart() + " to: " + getEnd() + ")";
-        }
-
-        public String getStart() {
-            return !(datestart == null) ? datestart.format(DateTimeFormatter.ofPattern("dd MMM, yyyy")) : start;
-        }
-
-        public String getEnd() {
-            return !(dateend == null) ? dateend.format(DateTimeFormatter.ofPattern("dd MMM, yyyy")) : end;
-        }
-
-        @Override
-        public String creationString() {
-            StringBuilder list = new StringBuilder();
-            if (this.isDone) {
-                list.append("d ");
-            }
-            list.append("event ").append(this.getDescription()).append(" /from ").append(start).append(" /to ").append(end);
-
-            return list.toString();
-        }
-    }
-
-    //used for loading and saving lists.
-    public static class Storage {
-        public static void saveList(TaskList list) throws IOException {
-            File folder = new File("data");
-            if (!folder.exists()) {
-                if (!folder.mkdirs()) {
-                    throw new IOException("folder creation broke");
-                }
-            }
-            File f = new File("data/list.txt");
-            if (!f.exists()) {
-                f.createNewFile();
-            }
-            FileWriter fw = new FileWriter("data/list.txt");
-            fw.write(list.listData());
-            fw.close();
-        }
-
-        //this just creates an empty list and saves it, to empty anything that used to be there
-        public static void clearSave() throws IOException {
-            TaskList list = new TaskList();
-            Storage.saveList(list);
-        }
-
-        //loads a list saved to the file location. It also checks if the file exists or if it is corrupted
-        public static TaskList loadList(String path) {
-            File f = new File(path);
-            TaskList list = new TaskList();
-            if (f.exists()) {
-                try {
-                    list.loadTask(f);
-                    UI.showMessage(list.returnList());
-                } catch (FileNotFoundException e) {
-                    UI.showMessage("this should never be seen");
-                } catch (InvalidFormatException e) {
-                    UI.showMessage("file corrupted, go see if theres anything salvagable because upon next list mod its all going to be overwritten");
-                }
-            }
-            return list;
-        }
     }
 }
