+++ src/main/java/doot/TodoTask.java
@@ -1,0 +1,33 @@
+package doot;
+
+//the subclass of task, specifically for the todo tasks
+public class TodoTask extends Task {
+    public TodoTask(String description) {
+        super(description);
+        type = Type.T;
+    }
+
+    //determines type of task, like T for todo
+    @Override
+    public String getType() {
+        return "[" +type.name() +"]";
+    }
+
+    //returns details for todo objects
+    @Override
+    public String getDetails() {
+        return this.getType() + this.getStatusIcon() + " " + this.getDescription();
+    }
+
+    //the string necessary to make this object again
+    @Override
+    public String creationString() {
+        StringBuilder list = new StringBuilder();
+        if (this.isDone) {
+            list.append("d ");
+        }
+        list.append("todo ").append(this.getDescription());
+
+        return list.toString();
+    }
+}
