+++ src/main/java/doot/Type.java
@@ -1,0 +1,3 @@
+package doot;
+
+public enum Type {T, D, E}
